+++ build_runner/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 2.7.0-wip
 
+- Performance: builders can choose to run only when "triggered". A builder runs
+  only if triggered if the option `run_only_if_triggered` is `true`. Triggers
+  are configured in new a top-level section of `build.yaml` called `triggers`.
+  See [the `build_config` docs](https://pub.dev/packages/build_config#triggers)
+  for more information.
 - Remove interactive prompts for whether to delete files.
 - Ignore `-d` flag: always delete files as if `-d` was passed.